backend/requirements.txt
flask
flask-sqlalchemy
uuid
jwt
flake8